メインViewの小数値表記のバグを修正

+++ app/src/main/java/com/t_robop/yuusuke/a01_spica_android/model/ScriptModel.java
@@ -14,7 +14,7 @@ public class ScriptModel extends BaseObservable {
     //スピードの三段階値
     private int speed = 2;
     //ブロック毎の値
-    private double value = 2;
+    private float value = 2;
     private boolean isInLoop=false;
 
     public ScriptModel() {
@@ -92,7 +92,7 @@ public class ScriptModel extends BaseObservable {
         this.speed = speed;
     }
 
-    public double getValue() {
+    public float getValue() {
         return value;
     }
